// apps/app/src/components/design-system/research-ui/documents/reference-prototype/types/documents.ts
export interface DocumentOneProps {
  header?: {
    title?: string
    subtitle?: string
    reference?: string
    classification?: string
  }
  metadata?: {
    justeret?: string
    sektm?: string
    udkstra?: string
    hjhabo?: string
    location?: string
  }
  footer?: {
    leftColumn?: {
      stehat?: string
      studio?: string
      items?: string[]
    }
    rightColumn?: {
      dseys?: string
      astso?: string
    }
  }
}